+++ ctek_njord_sim/run.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/with-contenv bashio
 # shellcheck shell=bash
-bashio::log.info "Starting CTEK Njord Load Balancer ..."
+bashio::log.info "Starting CtekNjorder ..."
 exec python3 -m app.main